5674: move shared files to $lib
