+++ widgets/src/lib/Icons/CheckMark.svelte
