+++ widgets/src/lib/Icons/Clock.svelte
