+++ widgets/src/lib/Icons/Close.svelte
