+++ widgets/src/lib/enums.ts
